Adding the additional auth0 API support (for create, updating emails, updating passwords, validating emails, etcetera)
Changed the gem to take a hash of options.  Added options to specify the attribute for the email attribute, password attribute, email_verified attribute, and the database connection name.  These all have default values.

Changed the sync_attr_with_auth0 method to be called after update instead of after save.  I added a method to search for the email address after_validation to ensure it's unique in auth0.  I also added a method to create the user in auth0 that fires after_create.  I also modified the sync_attr_with_auth0 methods to support synchronizing the email and password attributes, which have to be passed in separate API calls.

I added support for setting non-database attributes on the model to suppress any of the callbacks in the gem from firing.

I updated the spec to test the new functionality.

I updated the README to contain some information about how to use the gem.