Fixing issues communicating to auth0
Added auth0 and jwt as gem dependencies.

Updating the auth0 module to handle creating the JWT string and the connection object to the auth0 api.

Updated the model module in several ways.  The options are now kept in a single class attribute hash.  Several additional options were added, such as the name, given_name, and family_name.  All three callback methods were updated to use the auth0 gem.  The validate uses v1 of the api due to it not working on v2.  The create and sync methods both use v2 of the api.  The create will still set the email and a default password if the column does not exist on the model, but the sync will not touch either attribute in auth0.

The sync will always push all of the data, including the user's name into the auth0 app_metadata, and any additional information into the auth0 user_metadata.

I abstracted the user_metadata collection into a single method.  It will not include any information that is considered part of the app_metadata (like email, password, family_name, etc).