Case 18970 - Updating the sync gem to allow a password change to be pushed to auth0 on update of the user
Added an option to specify a verify_password_att which will tell auth0 if it needs to send a verification email or not.  If the attribute doesn't exist, it defaults to true.

Added the password and verify_password attributes to the arguments for the auth0 call if the password attribute should be sync'd and it was changed.