Case 18970 - Fixing a bug with updating the password
Chances are, if the password WAS a field on the database, it's being stored as an encrypted value.  When that happens, you can't test if password is being changed because it would be passing a hash, which is not useful for auth0.

If the password was NOT a field on the database, there is no "_changed?" method for the attribute, so we can't use that anyway.

So, we are going with the plan where the password attribute is only set on update if it is being changed.  So if you don't use a password field on the user, the attribute would remain nil for other changes.  If the password was a field on the user, you would be setting a different attribute so you could pass the plain-text version to auth0.

I implemented that change into the gem.