Case 19390 - Fixing an issue where if the password is the only change, the auth0_dirty? method comes up false
If the password attribute is not a DB attribute (which it really can't be unless you're storing plain text passwords), there is no "_changed?" method for it.  Because of that, the existing auth0_dirty? method doesn't capture when it is changed, and suppresses the sync.

I added a method called "auth0_user_password_changed?" that will do the necessary tests to see if the password was set/changed.  I added this call to the update method, replacing the original check.  I also added this call to the auth0_dirty? method to force it to return true if the password was set/changed.