Case 27716 - Making some major changes and refactoring to the gem
Updated the configuration so a config block can be used. Any setting on the model will override the config block.

Changed the default for the Auth0 namespace setting to look at 'AUTH0_NAMESPACE' instead of 'AUTH0_DOMAIN'.

Changed the attribute settings to use non-abbreviated variable names.

Updated the usage to include the fields to sync separate from the rest of the config.

A bunch of refactoring of methods into multiple modules to hopefully improve future maintenance (Configuration, Validation, Sync, etcetera).

Updated the sync so on create and on update, it determines if the sync to Auth0 should be a create or patch separately (to resolve issues with data inconsistency).

Updated the Auth0 update to always submit passwords and emails when they are changed, even if they are not fields to sync (that shouldn't be a requirement).