Unbind from json 1.8.1
Using this to help with implementing BloodRelay. Since it is a newer
app it was already on the 2.x series of the JSON gem.

Technically I could downgrade to 1.8.1 but then I would get warnings
due to the Ruby 2.4 integer unification. Since this version was pinned
with the initial commit I assume that pin is out-of-date and using a
newer version will work.

Ran the tests and they still all pass.

Went ahead and bumped the Gemfile.lock (bundle update). Generally I
don't think Gemfile.lock files are checked into repos for libraries
since the gemspec is all that is used.