Loosen JWT dependency
Had a hard dependency on 1.5.0. I was using 1.5.6 on BloodRelay. Tried
downgrading to match this gem but got some failures relating to not
verifying the JWT issuer.

Going to JWT 2.x is probably too big of a jump (don't think that is
compat with omniauth) but at least allowing a patch-level bump seems
reasonable and it fixes the failures.