Remove recursion possibility and reduce SQL queries
Previously if `sync_with_auth0_on_update=` was not defined it would
enter into a loop as setting that to value is the terminating condition.

Rather than go through the entire validation, callbacks, etc associated
with `save` which can trigger this sort of loop going with direct SQL
update of the column. This removes the need to disable syncing on the
object.

Also small optimization of only updating the column if it has changed.