Fix tests for saving `uid` from Auth0
In 3cd335b96ea3f6f3c72d4038fb949c5e7ab3282a I changed how the `uid`
is saved to remove the chance of a infinate loop as well as have lower
overhead. Unfortunately this change wasn't transparent to the test due
to the heavy mocking. This updates the test mocks to match the impl.