Rails 5.2+ compat
This gems determines if data needs to be pushed up to Auth0 by looking
at what attributes have changed. Unfortunately the "dirty" functionality
has changed with Rails 5.2 so that in the "after" save callback the
changed information is now reset.

A new set of methods have been created to access what was just saved.
I think this differs from the `previous_changes` list that Rails already
has although I'm vague on exactly how. A good article on the topic can
be found at:

https://www.levups.com/en/blog/2017/undocumented-dirty-attributes-activerecord-changes-rails51.html

We want to support the new API but since this gem is used on apps using
older versions of Rails I think we also want to support the old way
if the new way is not available. Once those apps are updated we might be
able to drop that support.

Alternatively we could just support the new way and older apps can only
use older versions of this gem, but that forces us to upgrade to the
latest Rails if they want fixes to this gem (or we have to backport
the fixes to old version of this gem). Since supporting both APIs is
trivial enough, it seemed best to me to just support both for now.

Renamed some methods internal to this gem for clarity:

- auth0_user_saved_changes_to_email? - Was `auth0_user_email_changed?`.
  Since this is targeting the new API it seemed best to reflect that in
  the name by using the same pattern.
- auth0_user_saved_changes_to_password? - Was
  `auth0_user_password_changed?`. Same reason as previous.
- save_to_auth0_after_create - Was `save_to_auth0_on_create`. To make
  it clear this is running in an `after` hook so we know what "dirty"
  methods we should be using.
- save_to_auth0_after_update - Was `save_to_auth0_on_update`. Same
  reasoning as previous.
- auth0_saved_changes_dirty? - Was `auth0_dirty?`. Make it clear that it
  is checking the saved changes dirty API and should only be used in
  an `after` callback.