Case 59567 - Upgrade to search engine v3
We are still using v2 of the management API but within that we are
switching to v3 of the search engine. As of June 30, 2019 v2 will be no
longer available.

Although this specific query is compatible with v3, specifying it will
prevent a warning message from showing in the Auth0 logs so we can
audit that we are no longer using v2 anywhere else and we are fully
upgraded to v3.