Case 59567 - Make search abide by connection
Previously it was searching all users which means it might pull a
user from a connection the app is not using. This never mattered before
because we just had one DB connection but now that BloodRelay has it's
own isolated user database the search needs to abide by the config.