Case 59567 - Allow search connection to be specified
We may create users in a specific database but we might attach to
users across multiple different connections, but not all connections.
This config allows you to specify what connections are used when finding
Auth0 users.

I debated if the default should be `connection_name` (so existing apps
would need to update to broaden their search) or if the default should
be all connections (keeping the historical behavior by default). Decided
the later was safer.