Case 187141 - Bump Dependencies
For BloodRelay I want to use a newer version of the Auth0 library.
URI.escape has been throwing deprecation warnings for a while but
now it's removed from Ruby 3 which means we need to be on a newer
version of Auth0 that no longer uses that method[1].

Had to decide if we want to stay on the 4.x series of Auth0 or the
5.x series due to a breaking change in `create_user` (or have our
library detect the version and call with the right arguments)[2]. I
opted to require v5 since it is better tested against recent versions
of Ruby[3].

Upgrading the auth0 gem required moving to the 2.2.x series of the JWT
library. The method doing the base64 encoding of the JWT signature has
moved. Also the `typ` header is no longer generated[4]. It seems it is
not required by the spec but Auth0 does require it so adding it as
a parameter manually. The algorithm param is the same as the default.
I just had to specify it to pass in header options.

Other than these library upgrades and the code changes caused by the
upgrades this commit includes no change in functionality. The test
suite passes but since it's heavily mocked I'll also note I have already
tested this against an upgraded version of BR to verify it really
doesn't break anything (this is how the `typ` thing was discovered).

1. https://github.com/auth0/ruby-auth0/issues/202
2. https://github.com/auth0/ruby-auth0/pull/244
3. https://github.com/auth0/ruby-auth0/pull/256
4. https://github.com/jwt/ruby-jwt/issues/233